pip-requirements.txt
mock==1.0.1
mockito==0.5.2
nose==1.3.0
wsgiref==0.1.2
dispatcher==1.0
mock==1.0.1
mockito==0.5.2
nose==1.3.0
wsgiref==0.1.2
